Migrate crm-web-api module; wrap methods in service
This migrates the full consultant-created crm api from old repos and uses the service to wrap its methods for isolation.
