+++ server/src/_utils/crm-web-api.ts
@@ -1,0 +1,144 @@
+import * as zlib from 'zlib';
+import * as request from 'request';
+import { ADAL } from './adal';
+
+export const CRMWebAPI = {
+  webAPIurl: 'process.env.webAPIurl',
+  CRMUrl: 'process.env.CRMUrl',
+  host() {
+    return `${this.CRMUrl}${this.webAPIurl}`;
+  },
+
+  dateReviver: function (key, value) {
+    if (typeof value === 'string') {
+      // YYYY-MM-DDTHH:mm:ss.sssZ => parsed as UTC
+      // YYYY-MM-DD => parsed as local date
+
+      if (value != "") {
+        const a = /^(\d{4})-(\d{2})-(\d{2})T(\d{2}):(\d{2}):(\d{2}(?:\.\d*)?)Z$/.exec(value);
+
+        if (a) {
+          const s = parseInt(a[6]);
+          const ms = Number(a[6]) * 1000 - s * 1000;
+          return new Date(Date.UTC(parseInt(a[1]), parseInt(a[2]) - 1, parseInt(a[3]), parseInt(a[4]), parseInt(a[5]), s, ms));
+        }
+
+        const b = /^(\d{4})-(\d{2})-(\d{2})$/.exec(value);
+
+        if (b) {
+          return new Date(parseInt(b[1]), parseInt(b[2]) - 1, parseInt(b[3]), 0, 0, 0, 0);
+        }
+      }
+    }
+
+    return value;
+  },
+
+  parseErrorMessage: function (json) {
+    if (json) {
+      if (json.error) {
+        return json.error;
+      }
+      if (json._error) {
+        return json._error;
+      }
+    }
+    return "Error";
+  },
+
+  // this provides the formatted values but doesn't do it for top level
+  // TODO: where should this happen? 
+  fixLongODataAnnotations: function (dataObj) {
+    return dataObj;
+  },
+
+  get: async function (query, maxPageSize = 100, headers= {}): Promise<any> {
+    //  get token
+    const JWToken = await ADAL.acquireToken();
+    const PreferHeaderOptions = [
+      'odata.include-annotations="OData.Community.Display.V1.FormattedValue"',
+      'odata.include-annotations="*"',
+      `odata.maxpagesize=${maxPageSize}`,
+    ].join(',');
+
+    const options = {
+      url: `${this.host() + query}`,
+      headers: {
+        'Accept-Encoding': 'gzip, deflate',
+        'Content-Type': 'application/json; charset=utf-8',
+        Authorization: `Bearer ${JWToken}`,
+        'OData-MaxVersion': '4.0',
+        'OData-Version': '4.0',
+        Accept: 'application/json',
+        Prefer: PreferHeaderOptions,
+        ...headers
+      },
+      encoding: null,
+    };
+
+    return new Promise((resolve, reject) => {
+      request.get(options, (error, response, body) => {
+        if (error) console.log(error);
+
+        const encoding = response.headers['content-encoding'];
+
+        if (!error && response.statusCode === 200) {
+          // If response is gzip, unzip first
+
+          const parseResponse = jsonText => {
+            const json_string = jsonText.toString('utf-8');
+
+            var result = JSON.parse(json_string, this.dateReviver);
+            if (result["@odata.context"].indexOf("/$entity") >= 0) {
+                // retrieve single
+                result = this.fixLongODataAnnotations(result);
+            }
+            else if (result.value ) {
+                // retrieve multiple
+                var array = [];
+                for (var i = 0; i < result.value.length; i++) {
+                  array.push(this.fixLongODataAnnotations(result.value[i]));
+                }
+                result.value = array;
+            }
+            resolve(result);
+          };
+
+          if (encoding && encoding.indexOf('gzip') >= 0) {
+            zlib.gunzip(body, (err, dezipped) => {
+                parseResponse(dezipped);
+            });
+          }
+          else {
+            parseResponse(body);
+          }
+        } else {
+          const parseError = jsonText => {
+            // Bug: sometimes CRM returns 'object reference' error
+            // Fix: if we retry error will not show again
+            const json_string = jsonText.toString('utf-8');
+            const result = JSON.parse(json_string, this.dateReviver);
+            const err = this.parseErrorMessage(result);
+
+            if (err == "Object reference not set to an instance of an object.") {
+              this.get(query, maxPageSize, options)
+                .then(
+                  resolve, reject
+                );
+            } else {
+              reject(err);
+            }
+          };
+
+          if (encoding && encoding.indexOf('gzip') >= 0) {
+            zlib.gunzip(body, (err, dezipped) => {
+              parseError(dezipped);
+            });
+          } else {
+            parseError(body);
+          }
+        }
+      });
+    });
+  },
+};
